requirements.txt
Flask==3.1.0
Flask-JWT-Extended==4.7.1
Flask-CORS==5.0.1
Flask-Limiter==3.9.0
flask-talisman==1.1.0
Flask-Migrate==4.1.0
Flask-WTF==1.2.2
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.36
bcrypt==4.2.1
python-dotenv==1.0.1
gunicorn==23.0.0
# psycopg2-binary==2.9.10  # Uncomment for PostgreSQL production deployment
email-validator==2.2.0
